requirements.txt
aiosqlite==0.22.0
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.115.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
openpyxl==3.1.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-telegram-bot==21.10
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.0
